+++ src/main/java/ru/spbau/mit/characters/mobs/Mob.java
@@ -1,4 +1,7 @@
-package ru.spbau.mit.characters;
+package ru.spbau.mit.characters.mobs;
+
+import ru.spbau.mit.characters.Character;
+import ru.spbau.mit.characters.Stats;
 
 public abstract class Mob extends Character {
     public Mob(char icon, Stats baseStats) {
