+++ src/main/java/ru/spbau/mit/characters/mobs/MobType.java
@@ -1,4 +1,4 @@
-package ru.spbau.mit.characters;
+package ru.spbau.mit.characters.mobs;
 
 public enum MobType {
     CREEP,
