+++ src/main/java/ru/spbau/mit/characters/mobs/MobsAI.java
@@ -1,0 +1,4 @@
+package ru.spbau.mit.characters.mobs;
+
+public class MobsAI {
+}
