--- src/main/java/ru/spbau/mit/events/Action.java
@@ -1,9 +1,0 @@
-package ru.spbau.mit.events;
-
-import ru.spbau.mit.core.GameState;
-
-import java.io.IOException;
-
-public interface Action {
-    void doAction(GameState gameState) throws IOException;
-}
